PinRepository.aggregateMonthYearCounts() // what an achievement :)

+++ life2-api/life2-api-artifact/src/main/java/w/whateva/life2/api/artifact/dto/ApiArtifactCount.java
@@ -9,5 +9,5 @@ public class ApiArtifactCount {
 
     Integer year;
     Integer month;
-    Integer count;
+    Double count;
 }
