+++ life2-data/life2-data-pin/src/main/java/w/whateva/life2/data/pin/domain/PinMonthYearCount.java
@@ -9,5 +9,5 @@ public class PinMonthYearCount {
 
     private Integer month;
     private Integer year;
-    private Integer count;
+    private Double count;
 }
